removed "DayTourForm.tsx" and merged into Create page
